Replace `Int64` by `Int`

+++ test/test_normal_form_game.jl
@@ -40,7 +40,7 @@ using CDDLib
     end
 
     @testset "Player with 2 opponents" begin
-        payoffs_2opponents = Array{Int64}(undef, 2, 2, 2)
+        payoffs_2opponents = Array{Int}(undef, 2, 2, 2)
         payoffs_2opponents[:, 1, 1] = [3, 1]
         payoffs_2opponents[:, 1, 2] = [6, 0]
         payoffs_2opponents[:, 2, 1] = [4, 5]
@@ -95,7 +95,7 @@ using CDDLib
     end
 
     @testset "asymmetric NormalFormGame with 3 players" begin
-        payoffs_2opponents = Array{Int64}(undef, 2, 2, 2)
+        payoffs_2opponents = Array{Int}(undef, 2, 2, 2)
         payoffs_2opponents[:, 1, 1] = [3, 1]
         payoffs_2opponents[:, 1, 2] = [6, 0]
         payoffs_2opponents[:, 2, 1] = [4, 5]
@@ -240,7 +240,7 @@ using CDDLib
         equal_po_p1_bimatrix[2, 1, :] = [1, 1]
         equal_po_p1_bimatrix[2, 2, :] = [1, -1]
 
-        three_p_equal_po_array = Array{Int64}(undef, 2, 2, 2)
+        three_p_equal_po_array = Array{Int}(undef, 2, 2, 2)
         three_p_equal_po_array[:, :, 1] = [2 0; 0 2]
         three_p_equal_po_array[:, :, 2] = [2 0; 0 2]
